refactor(test): JavaApplication 테스트 코드를 Spring Context를 활용하여 대체
